Refactor DB conn handling
db.SingleExec was created to reduce boilerplate for simple scrapes,
but I did not think about the overhead of opening a bunch of DB
conns in tight loops. Whoops.

This commit expands all of this back out such that the CLI is
responsible for passing a *pgx.Conn in to the scraper, allowing
for re-use.